✨ add links to text
