+++ client/components/videoplayer/helpers/chapters.ts
@@ -1,3 +1,5 @@
+import { getSecondsFromTimestamp } from 'shared';
+
 export const parseChapters = (description: string, videoLength: number) => {
   if (description.match(/0+:0+/gi)) {
     const timestampMatches = description.matchAll(/((\d{1,2}:)?\d{1,2}:\d{1,2})\s?-?\s?(.*)/gi);
@@ -28,22 +30,4 @@ export const parseChapters = (description: string, videoLength: number) => {
       };
     });
   }
-};
-
-const getSecondsFromTimestamp = (timestamp: string) => {
-  const timeStrings = timestamp.split(':');
-  const reverseTimeStrings = timeStrings.reverse();
-  let seconds = 0;
-
-  if (reverseTimeStrings.length > 0) {
-    seconds += parseInt(reverseTimeStrings[0]);
-  }
-  if (reverseTimeStrings.length > 1) {
-    seconds += parseInt(reverseTimeStrings[1]) * 60;
-  }
-  if (reverseTimeStrings.length > 2) {
-    seconds += parseInt(reverseTimeStrings[2]) * 3600;
-  }
-
-  return seconds;
 };
