+++ shared/index.ts
@@ -2,5 +2,12 @@ import { AuthorThumbnailDto } from './dto/video/author-thumbnail.dto';
 import { RecommendedVideoDto } from './dto/video/recommended-video.dto';
 import { VideoThumbnailDto } from './dto/video/video-thumbnail.dto';
 import { VideoDto } from './dto/video/video.dto';
+import { getSecondsFromTimestamp } from './util';
 
-export { AuthorThumbnailDto, RecommendedVideoDto, VideoThumbnailDto, VideoDto };
+export {
+  AuthorThumbnailDto,
+  RecommendedVideoDto,
+  VideoThumbnailDto,
+  VideoDto,
+  getSecondsFromTimestamp
+};
